Bring in changes from the 2024 lecture. (#44)
For the 2024 iteration of our lecture, we made some significant improvements to the mini Internet, which we now port back to the main branch.

Most notably:
- We removed the Open vSwitch connecting the containers and replaced it with direct virtual links. Adding links to the virtual switch scaled very badly to large topologies, resulting in multi-hour setup times. With the new implementation, setup takes at most tens of minutes, even for large topologies. Already "battle-tested" the new implementation earlier this year with over 100 students.
- Rewrote the restart scripts to work with the new implementation.
- Added a `HISTORY` service that starts with the mini-Internet and collects regular snapshots of all configs. This has two main purposes:
  - Even if students forget to save their config, we have a snapshot.
  - One can create a neat gif in the end that shows how connectivity evolved over time. The scripts to create this gif are in `utils/history_gif`
  - It can be disabled if space is a concern, but we found it to take a few hundred MB max, even after a long time.
- Added a `variables.sh` file in config. All configurable parameters have been removed from the scripts in `setup` and moved into `variables.sh` to make it easier to configure the Mini-Internet services in one place. This includes the webserver, matrix, and the new history service.

---------

+++ platform/cleanup/cleanup.sh
@@ -3,6 +3,7 @@
 # remove all container, bridges and temporary files
 # will only remove the containers, bridges defined in ../config/
 
+# set -x
 set -o errexit
 set -o pipefail
 set -o nounset
@@ -12,13 +13,36 @@ if [ "$#" != 1 ]; then
   exit 1
 fi
 
+# root privilege check in case this script is directly executed
+if (($UID != 0)); then
+    echo "$0 needs to be run as root"
+    exit 1
+fi
+
 DIRECTORY="$1"
 
 # kill all container
 ./cleanup/container_cleanup.sh "${DIRECTORY}"
 
 # remove all container & restart docker
+# remove all stopped containers, unused networks, dangling images and unused caches
+# -f: no confirmation
 docker system prune -f
+
+# find all namespaces with dangling symbolic links and remove them
+find /var/run/netns -xtype l -delete
+
+# # clear stale ovs interfaces on the server
+interface_list=$(ip link | grep -E '(_c@|vpn|_l|_a|_b|veth)' | awk -F': ' '{print $2}' | cut -d'@' -f1 || true) # ignore errors
+
+# Check if interface_list is not empty
+if [ -n "$interface_list" ]; then
+    echo "$interface_list" | while read -r interface; do
+        # Delete the interface, ignoring any errors
+        ip link delete "$interface" || true
+    done
+fi
+
 
 echo -n "ovs-vsctl " > ovs_command.txt
 
@@ -33,6 +57,7 @@ echo -n "ovs-vsctl " > ovs_command.txt
 ./cleanup/vpn_cleanup.sh "${DIRECTORY}"
 
 
+# ensure any failure in the executed commands does not stop the script due to errexit
 bash  < ovs_command.txt || true
 rm -f ovs_command.txt
